SentimentAnalysis/requirements.txt
absl-py==0.7.1
astor==0.7.1
boto==2.49.0
boto3==1.9.142
botocore==1.12.142
certifi==2019.3.9
chardet==3.0.4
Click==7.0
cycler==0.10.0
docutils==0.14
Flask==1.0.3
Flask-Login==0.4.1
Flask-LoginManager==1.1.6
gast==0.2.2
gensim==3.7.2
glove-python==0.1.0
grpcio==1.20.1
h5py==2.9.0
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.1
jmespath==0.9.4
joblib==0.13.2
Keras==2.2.4
Keras-Applications==1.0.7
Keras-Preprocessing==1.0.9
kiwisolver==1.1.0
Markdown==3.1
MarkupSafe==1.1.1
matplotlib==3.0.3
mock==3.0.4
nltk==3.4
numpy==1.16.2
protobuf==3.7.1
pyparsing==2.4.1.1
PyStemmer==1.3.0
python-dateutil==2.8.0
PyYAML==5.1
requests==2.21.0
s3transfer==0.2.0
scikit-learn==0.21.2
scipy==1.2.1
singledispatch==3.4.0.3
six==1.12.0
smart-open==1.8.3
tensorboard==1.13.1
tensorflow==1.13.1
tensorflow-estimator==1.13.0
termcolor==1.1.0
urllib3==1.24.3
Werkzeug==0.15.2
